backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
asyncpg==0.29.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose==3.3.0
passlib==1.7.4
cryptography==43.0.1
httpx==0.27.2
email-validator==2.2.0
